improve regular file handling
The --dereference= option expects symlink as argument and
now tha symlink can point to a regular file, effectively
copying it into the current directory.
The --prune= option can now take a regular file path as
its argument. Prior to this it expected a directory path
as its argument,
Also fix issue in which multiple clones to the same
destination would not properly overwrite regular files.
When a non-root user copies a regular file that in
SPATH had read-only permissions for that user, then
those permissions need to be expanded to add user
write permission in the output (i.e. under DPATH).



git-svn-id: https://svn.bingwo.ca/repos/clone_pseudo_fs/trunk@16 b16cee52-b522-4eb8-934a-94b5746ee385